requirements.txt
# Core LangChain and LLM dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-classic>=0.0.1
langchain-openai>=0.0.5
openai>=1.0.0

# Environment management
python-dotenv>=1.0.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Speech-to-Text (Whisper)
openai-whisper>=20231117
# Deepgram for diarization
deepgram-sdk>=3.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0

# Text-to-Speech for dummy audio generation
gTTS>=2.3.0

# Audio processing
pydub>=0.25.0

# Utilities
requests>=2.31.0
